\name{plot-methods}
\docType{methods}
\title{Methods to plot the verification plot of the obtained optimal design.}
\description{
Function MOPT, ceff1, ceff2, and Deff return objects of class OPT. The plot-methods plot the verification plot of the multiple-objective optimal design (from function MOPT),
the D-optimal design (from function Deff), the c-optimal design for estimating the ED50 (from function ceff1), and the c-optimal design for estimating the MED (from function ceff2).
}
\section{Methods}{
\describe{

\item{\code{signature(x = "OPT")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}

