\name{summary-methods}
\docType{methods}
\title{Methods to print SW objects or the optimal design and the efficiency from OPT objects.}
\description{
Function MOPT, ceff1, ceff2, and Deff return objects of class OPT. Function S.Weight returns objects of class SW. 
The summary-methods print the multiple-objective optimal design (from function MOPT); the c-optimal design and c-efficiency for estimating the ED50 (from function ceff1);
the c-optimal design and c-efficiency for estimating the MED (from function ceff2); the D-optimal design and D-efficiency (from function Deff); the matrix of the optimal weights
and the first and second derivative of the optimality criterion w.r.t weights (from function S.Weight).   
}
\section{Methods}{
\describe{

\item{\code{signature(object = "OPT")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SW")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
